Extract worktree, rpc_client, and util crates
