+++ worktree/Cargo.toml
@@ -1,0 +1,61 @@
+[package]
+name = "worktree"
+version = "0.1.0"
+edition = "2018"
+
+[features]
+test-support = []
+
+[dependencies]
+anyhow = "1.0.38"
+# async-recursion = "0.3"
+async-trait = "0.1"
+# async-tungstenite = { version = "0.14", features = ["async-tls"] }
+buffer = { path = "../buffer" }
+clock = { path = "../clock" }
+# crossbeam-channel = "0.5.0"
+# ctor = "0.1.20"
+# dirs = "3.0"
+# easy-parallel = "3.1.0"
+fsevent = { path = "../fsevent" }
+futures = "0.3"
+fuzzy = { path = "../fuzzy" }
+gpui = { path = "../gpui" }
+# http-auth-basic = "0.1.3"
+ignore = "0.4"
+# image = "0.23"
+# indexmap = "1.6.2"
+lazy_static = "1.4.0"
+libc = "0.2"
+log = "0.4"
+# log-panics = { version = "2.0", features = ["with-backtrace"] }
+# num_cpus = "1.13.0"
+parking_lot = "0.11.1"
+postage = { version = "0.4.1", features = ["futures-traits"] }
+rpc_client = { path = "../rpc_client" }
+# rand = "0.8.3"
+# rsa = "0.4"
+# rust-embed = { version = "6.2", features = ["include-exclude"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = { version = "1.0.64", features = ["preserve_order"] }
+# serde_path_to_error = "0.1.4"
+# simplelog = "0.9"
+# smallvec = { version = "1.6", features = ["union"] }
+smol = "1.2.5"
+sum_tree = { path = "../sum_tree" }
+util = { path = "../util" }
+# surf = "2.2"
+# tempdir = { version = "0.3.7", optional = true }
+# thiserror = "1.0.29"
+# time = { version = "0.3" }
+# tiny_http = "0.8"
+toml = "0.5"
+# tree-sitter = "0.19.5"
+# tree-sitter-rust = "0.19.0"
+# url = "2.2"
+zrpc = { path = "../zrpc" }
+
+[dev-dependencies]
+rand = "0.8.3"
+tempdir = { version = "0.3.7" }
+util = { path = "../util", features = ["test-support"] }
