+++ worktree/src/ignore.rs
