+++ zed/src/chat_panel.rs
@@ -3,10 +3,7 @@ use std::sync::Arc;
 use crate::{
     channel::{Channel, ChannelEvent, ChannelList, ChannelMessage},
     editor::Editor,
-    rpc::{self, Client},
-    theme,
-    util::{ResultExt, TryFutureExt},
-    Settings,
+    theme, Settings,
 };
 use gpui::{
     action,
@@ -18,12 +15,14 @@ use gpui::{
     ViewContext, ViewHandle,
 };
 use postage::{prelude::Stream, watch};
+use rpc_client as rpc;
 use time::{OffsetDateTime, UtcOffset};
+use util::{ResultExt, TryFutureExt};
 
 const MESSAGE_LOADING_THRESHOLD: usize = 50;
 
 pub struct ChatPanel {
-    rpc: Arc<Client>,
+    rpc: Arc<rpc::Client>,
     channel_list: ModelHandle<ChannelList>,
     active_channel: Option<(ModelHandle<Channel>, Subscription)>,
     message_list: ListState,
@@ -48,7 +47,7 @@ pub fn init(cx: &mut MutableAppContext) {
 
 impl ChatPanel {
     pub fn new(
-        rpc: Arc<Client>,
+        rpc: Arc<rpc::Client>,
         channel_list: ModelHandle<ChannelList>,
         settings: watch::Receiver<Settings>,
         cx: &mut ViewContext<Self>,
